pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Agentic Service scaffold (Manager-Planner-Executor-Infrastructure)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-adk>=1.25.0",
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "opensearch-py>=2.8.0",
  "redis>=5.2.0",
  "httpx>=0.28.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "mypy>=1.14.0",
  "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
